requirements.txt
aiogram==3.25.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
python-dotenv==1.1.0
httpx==0.28.1
zeep==4.3.1
apscheduler==3.11.0
fastapi==0.115.6
uvicorn==0.34.0
redis==5.2.1
